created tariff controller

+++ Server/Models/Tariff.cs
@@ -8,6 +8,6 @@ public record Tariff
     public int? Id { get; init; }
     public string? Name { get; init; }
     public string? Description { get; init; }
-    public double? Price { get; init; }
+    public decimal? Price { get; init; }
     public int? Car { get; init; }
 }
